feat(pro:search): search btn triggers search change now
