+++ packages/pro/search/src/ProSearch.tsx
@@ -18,9 +18,11 @@ import { useCommonOverlayProps } from './composables/useCommonOverlayProps'
 import { useFocusedState } from './composables/useFocusedState'
 import { useSearchItems } from './composables/useSearchItem'
 import { useSearchItemErrors } from './composables/useSearchItemErrors'
-import { useSearchStates } from './composables/useSearchStates'
+import { tempSearchStateKey, useSearchStates } from './composables/useSearchStates'
+import { useSearchTrigger } from './composables/useSearchTrigger'
 import { useSearchValues } from './composables/useSearchValues'
 import SearchItemComp from './searchItem/SearchItem'
+import SearchItemTagComp from './searchItem/SearchItemTag'
 import { proSearchContext } from './token'
 import { type SearchItem, proSearchProps } from './types'
 import { renderIcon } from './utils/RenderIcon'
@@ -47,6 +49,7 @@ export default defineComponent({
       errors,
       dateConfig,
     )
+    const searchTriggerContext = useSearchTrigger()
     const elementRef = ref<HTMLElement | undefined>()
 
     const activeSegmentContext = useActiveSegment(
@@ -67,7 +70,7 @@ export default defineComponent({
     //TODO: fix ref(true)
     const currentZIndex = useZIndex(toRef(props, 'zIndex'), toRef(componentCommon, 'overlayZIndex'), ref(true))
 
-    const { initSearchStates, clearSearchState } = searchStateContext
+    const { initSearchStates, clearSearchState, getSearchStateByKey } = searchStateContext
     const { activeSegment } = activeSegmentContext
 
     watch(
@@ -99,8 +102,13 @@ export default defineComponent({
 
     expose({ focus, blur })
 
-    const handleSearchBtnClick = () => {
+    const { onSearchTrigger, triggerSearch } = searchTriggerContext
+    onSearchTrigger(() => {
       callEmit(props.onSearch, searchValues.value)
+    }, 'post')
+
+    const handleSearchBtnClick = () => {
+      triggerSearch()
     }
     const handleClearBtnClick = () => {
       clearSearchState()
@@ -121,13 +129,26 @@ export default defineComponent({
 
       ...searchStateContext,
       ...activeSegmentContext,
+      ...searchTriggerContext,
     })
 
     return () => {
       const prefixCls = mergedPrefixCls.value
 
       const overflowSlots = {
-        item: (item: SearchItem) => <SearchItemComp key={item.key} tagOnly={true} searchItem={item} />,
+        item: (item: SearchItem) => {
+          const searchState = getSearchStateByKey(item.key)!
+
+          const tagSegments = item.segments.map(segment => {
+            const segmentValue = searchState.segmentValues.find(sv => sv.name === segment.name)!
+            return {
+              name: segment.name,
+              input: segment.format(segmentValue?.value),
+            }
+          })
+
+          return <SearchItemTagComp key={item.key} itemKey={item.key} segments={tagSegments} error={item.error} />
+        },
         rest: (rest: SearchItem[]) => (
           <span class={`${prefixCls}-search-item ${prefixCls}-search-item-tag`}>
             {slots.overflowedLabel?.(rest) ?? `+ ${rest.length}`}
@@ -143,7 +164,7 @@ export default defineComponent({
                 v-show={!focused.value}
                 v-slots={overflowSlots}
                 prefixCls={prefixCls}
-                dataSource={searchItems.value}
+                dataSource={searchItems.value.filter(item => item.key !== tempSearchStateKey)}
                 getKey={item => item.key}
                 maxLabel={props.maxLabel}
               />
