+++ packages/pro/search/src/token.ts
@@ -7,6 +7,7 @@
 
 import type { ActiveSegmentContext } from './composables/useActiveSegment'
 import type { SearchStateContext } from './composables/useSearchStates'
+import type { SearchTriggerContext } from './composables/useSearchTrigger'
 import type { SegmentOverlayUpdateContext } from './composables/useSegmentOverlayUpdate'
 import type { SegmentStatesContext } from './composables/useSegmentStates'
 import type { ProSearchProps } from './types'
@@ -14,7 +15,7 @@ import type { ɵOverlayProps } from '@idux/components/_private/overlay'
 import type { ProSearchLocale } from '@idux/pro/locales'
 import type { ComputedRef, InjectionKey, Slots } from 'vue'
 
-export interface ProSearchContext extends SearchStateContext, ActiveSegmentContext {
+export interface ProSearchContext extends SearchStateContext, ActiveSegmentContext, SearchTriggerContext {
   props: ProSearchProps
   slots: Slots
   locale: ProSearchLocale
