+++ packages/pro/search/src/types/index.ts
@@ -1,0 +1,13 @@
+/**
+ * @license
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://github.com/IDuxFE/idux/blob/main/LICENSE
+ */
+
+export * from './segment'
+export * from './searchItem'
+export * from './searchValue'
+export * from './searchFields'
+export * from './panels'
+export * from './proSearch'
